feat: Update data fetching to handle list of loan items and display loan purpose
-   Modified `ApiService` to return a list of `LoanResponseItem` instead of `LoanResponse`.
-   Updated `LoanRepository` to work with a list of `LoanResponseItem` in `NetworkBoundResource`.
-   Modified `LoanDataMapper` to process a list of `LoanResponseItem` for domain mapping.
-   Updated `RemoteDataSource` to check if the list of loan response items is empty.
-   Updated `LoanListFragment` to observe the loan data, display the first loan's purpose in a toast, and set the `tvLoanList` text.
-   Updated `Loan` domain model to use `Double` for `loanAmount` and `collateralValue`.
-   Removed unused `LoanResponse` data class and added `LoanResponseItem`.

+++ core/src/main/java/com/ekachandra/core/domain/model/Loan.kt
@@ -6,14 +6,14 @@ import com.ekachandra.core.data.source.remote.response.RepaymentSchedule
 data class Loan(
     val borrowerName: String? = null,
     val borrowerEmail: String? = null,
-    val loanAmount: Int? = null,
+    val loanAmount: Double? = null,
     val interestRate: Double? = null,
     val term: Int? = null,
     val purpose: String? = null,
     val riskRating: String? = null,
     val creditScore: Int? = null,
     val collateralType: String? = null,
-    val collateralValue: Int? = null,
+    val collateralValue: Double? = null,
     val repaymentSchedule: RepaymentSchedule? = null,
     val documents: List<DocumentsItem?>? = null,
 )
